+++ src/routers/task.js
@@ -1,0 +1,80 @@
+const express = require("express");
+const router = new express.Router();
+const Task = require("../models/task")
+
+router.post("/tasks", async (req, res) => {
+  try {
+    const task = new Task(req.body);
+    await task.save();
+    res.status(201).send(task);
+  } catch (error) {
+    res.status(400).send(error);
+  }
+});
+
+router.get("/tasks", async (req, res) => {
+  try {
+    const tasks = await Task.find({});
+    res.status(202).send(tasks);
+  } catch (error) {
+    res.status(500).send(error);
+  }
+});
+
+router.get("/tasks/:id", async (req, res) => {
+  try {
+    const task = await Task.findById(req.params.id);
+    if (task) {
+      res.send(task);
+    } else {
+      res.status(404).send();
+    }
+  } catch (error) {
+    res.status(500).send();
+  }
+});
+
+router.patch("/tasks/:id", async (req, res) => {
+  const updates = Object.keys(req.body);
+  const allowedUpdates = ["description", "completed"];
+  const isValidOperation = updates.every((update) =>
+    allowedUpdates.includes(update)
+  );
+
+  try {
+    if (!isValidOperation) {
+      return res
+        .status(400)
+        .send({ error: "No properties allowed to update." });
+    }
+
+    const task = await Task.findByIdAndUpdate(req.params.id, req.body, {
+      new: true,
+      runValidators: true,
+    });
+
+    if (!task) {
+      return res.status(404).send();
+    }
+    return res.send(task);
+  } catch (error) {
+    console.error(error);
+    res.status(500).send(error);
+  }
+});
+
+router.delete("/tasks/:id", async (req, res) => {
+  try {
+    const task = await Task.findByIdAndDelete(req.params.id);
+    if (task) {
+      return res.send(task);
+    } else {
+      return res.status(404).send();
+    }
+  } catch (error) {
+    console.error(error);
+    res.status(500).send();
+  }
+});
+
+module.exports = router;
